// app/backend/src/Interfaces/Leaderboard/ILeaderboard.ts
export interface ILeaderboard {
  name: string;
  totalPoints: number;
  totalGames: number;
  totalVictories: number;
  totalDraws: number;
  totalLosses: number;
  goalsFavor: number;
  goalsOwn: number;
  goalsBalance: number;
  efficiency: number;
}

export type TeamGoals = 'homeTeamGoals' | 'awayTeamGoals';